Add model sync behavior for gmaps torque layer view too
Solved by extracting the implementation from previous commit to an
object that can be mixed into the two different implementations instead
of duplicating the logic.
