+++ grunt/tasks/_browserify-bundles.js
@@ -1,4 +1,5 @@
 var torqueSpecs = [
+  'test/spec/geo/gmaps/gmaps-torque-layer-view.spec.js',
   'test/spec/geo/leaflet/leaflet-torque-layer.spec.js',
   'test/spec/geo/ui/layer-selector-torque.spec.js'
 ];
